Added database / hibernate shtuff
I have started work for persisting stuff to a MySQL database.  The
database script to create the schema is in the databaseScripts folder.
Any sequential database scripts should be placed in a  new file in this
folder.  Also added support for Hibernate and started work on creating
a DataAccess class